+++ ExportCategories.lua
@@ -426,6 +426,7 @@ DF.ExportCategories = {
         "directBuffSortOrder",
         "directDebuffShowAll",
         "directDebuffFilterRaid",
+        "directDebuffFilterRaidInCombat",
         "directDebuffFilterCrowdControl",
         "directDebuffFilterImportant",
         "directDebuffSortOrder",